lang/funcs: Make nonsensitive more permissive
Calling the nonsensitive function with values which are not sensitive
would previously error. This restriction was added with the goal of
preventing confusingly redundant use of this function.

Unfortunately this breaks when using nonsensitive to reveal the value of
sensitive resource attributes. This is because the validate walk does
not (and cannot) mark attributes as sensitive based on the schema,
because the resource value itself is unknown.

This commit therefore removes this restriction, and adds a test case to
cover this specific scenario.